pyproject.toml
[tool.poetry]
name = "jpeg-filter-prototype"
version = "0.1.0"
description = ""
authors = ["Yusuke Kawamoto <novogrammer@gmail.com>"]
readme = "README.md"
packages = [{include = "jpeg_filter_prototype"}]

[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.8.0.74"
numpy = "^1.25.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
